nit; rename function
