+++ contents/index.ts
@@ -5,7 +5,7 @@ import catchHuluSubtitles from "~contents/catchHuluSubtitles"
 import { observeSection } from "~utils"
 import { USER_SETTINGS_DEFAULTS } from "~utils/constants"
 import { pollCachedTranslations } from "~utils/functions/cachedTranslations"
-import extractIdFromUrl from "~utils/functions/extractMovieFromNetflixUrl"
+import extractIdFromUrl from "~utils/functions/extractIdFromUrl"
 import initData from "~utils/functions/initData"
 import handleUrlChange from "~utils/handlers/handleUrlChange"
 import { waitForElement } from "~utils/index"
